# perf(storage): improve scanning with GLSN